bot/requirements.txt
python-telegram-bot[job-queue]
libtorrent
